Move footer code into an external file in _includes
This change is needed because the footer code needs to be duplicated in
two places:

1. A footer inside a Flex-box in the sidebar. Flex-box is used as an
alternative to the fixed footer. In fact,
`justify-content: space-between` is used to keep an empty space between
the header and the footer; but if the View-port gets too small, elements
will not overlap;

2. When the page width is less than 960px, the old footer tag will be
shown (with `position: static` property) and the new footer inside the
sidebar will be hidden.