Add a top and bottom padding to the sidebar
The main content of the page (section tag) is affected by the padding
and margin from the body tag (50px for the padding and the default 8px
margin). To keep the top line of the header aligned with the section
tag, a 58px top padding is needed.

The original footer has a bottom property equals to 50px. To recreate
this behavior, we use a bottom padding with the same size.