Use sidebar wrapper as a Flex-box container for header and footer
The idea behind this branch is to fix the overlapping header and footer
using a Flex-box.

`justify-content: space-between` is used because we have only two child
elements (header and footer) and we want the header to be at the very
top of the View-port, and the footer at the very bottom of the
View-port.

The sidebar has a fixed position without any `top` or `bottom`
properties. Its height is then equal to the total View-port height minus
the paddings. We have a 58px top padding and a 50px bottom padding.
The final value is calculated by the browser using the formula
(100vh - 108px).