Make download buttons fit inside the header container
The right border of the right button does not show up.

I believe that this is a regression caused by changes in this branch.

Is is probably caused by the CSS properties added to hide the scrollbar
but keep the scrolling functionality in the sidebar.

The downloads list width is 270px and it has 1px border. The sum is
272px. But the container which is the header tag offers a 270px width
size only. For this reason, the width is reduced from 270px to 268px.

li elements inside the ul downloads list have a fixed size of 89px.
The three elements can not fit anymore in the new container width of
268px. Their widths are then reduced to 88px to make them fit.